Move providers to new folder
