+++ client/src/app/hooks/useError.ts
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import { ErrorContext } from "../components/ErrorProvider";
+import { ErrorContext } from "../components/providers/ErrorProvider";
 
 export function useError() {
   const error = useContext(ErrorContext);
