+++ client/src/app/hooks/useWormhole.ts
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import { WormholeContext } from "../components/WormholeProvider";
+import { WormholeContext } from "../components/providers/WormholeProvider";
 
 export function useWormhole() {
   const wormhole = useContext(WormholeContext);
